config/: In devel, always use Unix sockets on local machine
This ensures that we'll always end up using peer authentication (that
is, postgres will just ask the socket for the name of the Unix user on
the other end).

It turns out that on some systems (e.g. Ubuntu 14.04) by default
localhost connections will go through TCP/IP, which doesn't work with
peer auth.