Add OAuth and automatic user creation.
The OAuth functionality is all provided by the martini-oauth
middleware.

This adds another middleware that detects the user_id in the session.
If none is found, it attempts to look up the user's email address
using the OAuth token, and map that to a Parsley user. If no user
exists (or there are no emails with any of the addresses from
google) a new user is created and the user is redirected to the
/signup page.