Add goose for migrations
Heroku wasn't happy with the migration strategy that I had developed.
This is more traditional, using a separate app for the migrations.

Note: It requires a custom version of gorm, which adds the gorm.Wrap
command.  I have requested it be pulled into the main repo but in
the meantime I will maintain it under github.com/jnfeinstein/gorm.