controllers/supplier: Create per-user views of this table
Whether a resource is global or in a per-user sandbox is not obvious. In
the future, we might want some fields (e.g. the supplier's
name/existence) to be public, and others (like contact info) stay
private, but if we want to sandbox a whole DB table this is how to do
it.

Possible future tweaks:

* We could have some globally-visible rows (ingredient info from the
FDA database, for example) in a generally sandboxed table by having a
special global UserId (e.g. 0).

* Another thing we might want to do is sandbox relations (e.g. even for
the shared ingredients, each user has different suppliers).