parsley: Add sub-router for Organization (with supplier routes)
I added a component to handle every URL under and including
"/parsley/organizations/:org_id". This component handles tacking on the
SecondaryNavBar and also includes its own contextual router (sub-router)
to pick the right component to render the body of the page. I added a
(mostly non-functional) Supplier component to get this to work.

In the process, I pulled almost all of the logic out of the
OrganizationComponent; that now only deals with display and editing of
whatever is in the organization prop.