Now we can use react-router-component to navigate and react-bootstrap wont do stupid shit.
react-router-component REALLY wants you to use their navigate function.
I found that it is exposed in their defaultEnvironment variable. I then
wrapped it in a helper and added it to the defaultProps for the bootstrap
NavItem. Now when you click on any NavItem, it will use the helper
to navigate.