Add .gitattributes file for better diffs
This makes --color-words do something a bit better. The cpp diff engine
is unfortunately not perfect for either .go or .js, but it's better than
the default tokenization into runs of non-whitespace characters.
Possibly there's a better solution out there for both languages, but I'm
too lazy right now to look for one.